{bin,pylib}/cmtconv: CMT tape audio ↔ data conversion
This is a Python library (cmtconv) and program (bin/cmtconv) that converts
WAV file recordings of cassette tape saves to various data formats and vice
versa. It currently supports the National/Panasonic JR-200 and the Hitachi
Basic Master Jr.

As well as the pytest unit tests, there is also a set of (noisy) functional
tests in pylib/cmtconv/Test. At least some of these tests can probably be
converted to unit tests; the remainder should be made quieter and we should
provide some way of ensuring that they're usually called from top-level
tests so that they don't get forgotten.

TODO: Some of the v4()-level logging is commented out because they generate
a lot of output and greatly slow the unit tests (because the unit test
framework captures that output). In order to uncomment these we should
actually change the test framework for the CMT unit tests not to print
debugging output (which is not shown anyway if the test passes) unless the
developer requests it.

HISTORICAL NOTE: This is a tweaked version of this commit from 8bitdev[1]
to fix dependencies (e.g., testmc.memimage is now binary.memimage). For
this and future commits you still can see the original history (which is
substantially the same) in that repo. Currently the pylib/cmtconv/Test
script doesn't run due to the lack of some other files that have not been
moved over from 8bitdev; this is fixed in future commits.

[1]: https://github.com/0cjs/8bitdev.git