tokentext: Rename class and implement setline() and text()
The class is renamed from TokenText to TokLines, since it deals only with
the higher-level structure of the lines (offset, line number and arbitrary
line data) and knows nothing about the contents of an individual line,
other than it may not have a 00 byte in it.

parsetext(), which builds the data structure from an initial binary dump of
the text data, is not yet implemented.