toklines: Lines _can_ contain $00 chars; rewrite parsing
It turns out that tokenized lines can contain $00 chars; one example
would be a (pre-RUN) line number after a GOTO, encoded as $0E $nn $nn,
where $nn $nn is a 16-bit unsigned int.

So now we have had to rewrite the parsing to work based on the
next-line addresses in the text, rather than trying to search for
termination characters. This unfortunately means that we can no longer
guess txttab (the start address of the program), since we can't find
the start of the second line without already knowing txttab. (Well, we
could try to use a hueristic, possibly involving detokenization/
parsing of line data, but that's too much for now.)