bastok.msx2: detokenize(), without non-ASCII char support
This has not been extensively tested yet, but several hundred lines
containing only ASCII characters from a BASIC game have been checked
against `SAVE "...",A` results and they match.

The biggest thing missing at this point is support for translation tables
for non-ASCII characters.

A class that takes parameters on initialization and has a single method to
do its job and return its result is not a great interface (a simple
function would be better), but this makes it easier to share internal
tokenizer state amongst all the various functions. Also, there will also
soon be further parameters to deal with the formatting of the
detokenization.