bastok.msx2: We now "generate" output; add function to retrieve output
We now refer to appending to the output as "generation," and rename the old
output() function to gen(). This frees up the output() name for a new
function that returns the generated output. both giving a slightly more
clear interface and letting the unit tests do less digging into the
internals of the class.

As well, we now generate the line number on reset(), rather than waiting
until generation is done and pasting it in front.