bastok.msx2: Clarify ASCII vs. MSX charset generation
If we're using a custom MSX character set that redefines the codes normally
the same as ASCII, we still want to use ASCII for the BASIC code itself. So
to help with this, we clarify when we are generating ASCII and when we are
generating Unicode substitutes for characters in the MSX character set (in
string constants and the like).

To do this we add a genasc() function that takes an ASCII code or a `str`
containing only ASCII characters, and generates an ASCII character or
characters to the output, and use this for all BASIC code. We also rename
nchar() to asc(), since it's used to consume and print only ASCII
characters.

This also centralises int ASCII code to ASCII character conversion in one
place, since it's now done by genasc() instead of at every generation call
point where we had an int containing the ASCII code.