bin/detok,bastok.msx2: Add binary mode and always use UTF-8 in trans mode
If we are translating characters to Unicode in the output, always use
UTF-8. This not only avoids throwing an exception if the current locale
doesn't support the Unicode characters we use, but also ensures that we get
consistent output regardless of locale because the output would typically
be committed as a "source" file.

Also, add the -b/--binary option to simply pass the MSX-BASIC-encoded bytes
straight through instead of converting them. This should produce output
exactly the same as what SAVE "file.bas",A would produce. This can be
useful for checking that the detokenizer is working correctly. (But this
does have limits; files can probably be generated that will confuse the
MSX-BASIC detokenzer, this detokenizer, or both.)

After adding this option the output was checked against a thousand-odd
lines of a BASIC game and the only difference that turned up was that this
program appended `!` to small floating point constants where the ,A save
did not.