bastok.msx2: Make INTERVAL a separate token, and rework token parsing
INTERVAL is actually tokenized as the INT and VAR tokens with ASCII letters
'ER' between them.¹ We now consider this whole sequence as one token and
switch to PEG-style parsing so we match it before INT, thus fixing both the
issue of the expander putting spaces in the middle of the keyword and
making this more sensible should this code be used for actual language-
level parsing at any point.

We had to deal with expansion of `INTERVALON` (which should get a space
before `ON`) and `ON INTERVAL=30` (which should not have a space before
around the `=`); the code for this by happy coincidence lets us get rid of
the trailing space problem we had with things like `PRINT:` as well.
Basically, we just look forward to determine if the next character is in
the small set that should not be prefixed by a space (EOL, ':', '(', and
'='), which seems to neatly solve that problem.

There are, however, still areas where expander incorrectly adds a trailing
space, such as when the final item in a DATA statement is empty, and where
it doesn't add a space where it would be good to have one, such as `ON
INTERVAL=15GOSUB 80`. Fixing things like this is probably not going to
happen because the desire for a space before the GOSUB (or after the
number) is dependent on the context of the GOSUB, which would mean we'd
have to start doing some real language parsing. It's easier just to let the
developers fix it by hand when they generate the output, since they do it
only once.

Various lower-level parsing functions are cleaned up, added and tested, and
the keyword() function is now replaced by the better-named token().
However, this is still not used for all tokens; we check for quite a number
tokens not in the table (such as integer constants) and even use different
processing for a few that are in the table (such as REM and DATA). Perhaps
at some point we can move towards having the table specify the parser
combinator for each entry so that we can make the parser itself more
generic.

We introduce a new ParseError to try to separate internal errors (e.g.,
trying to consume the wrong character when we already know what it is) from
a malformed token stream, though it's not clear how well this is really
working out.