bin/blines, bastok.blines: Turn expanded BASIC programs back to single lines
This deals with converting multiline back to single line and removing any
comments that may have been added. This may be useful as a separate tool
for systems where we don't have a retoknizer yet.

`msx2re` is not a good name; adding this seems to have exposed a deficiency
in the structure of our module naming that we will fix in future commits,
as a better structure becomes clear.