bastok.charset: Variable Unicode offsets for UTCharset
In some cases we're testing things where the underlying charset
conversion has already been tested and we'd like our strings to be
simpler and shorter.

This allows us to change the offset on the Unicode side from 0xF000
to, e.g., 0x0000, which is essentially a mapping of native charset
directly to the first 256 code points in Unicode, producing more
readable output from tests that are not concerned with testing
character set translation.