bastok.charset: native() now returns an int, not a bytes, and doc fix
From the comment, previously there was clearly some confusion that native()
should have been returning an MSX-BASIC _encoded_ sequence of bytes (e.g.,
'\x01\x45' for a Unicode character, rather than the MSX character set code
point (e.g., 0x05). Document that it's the latter, and change the return
type from `bytes` to `int` because that's more convenient for the parser
routines that do encoding, as we shall see later.

We also take this opportunity to add some Charset tests to charset.pt.