activate → pactivate: Update python virtualenv setup script
From: https://github.com/0cjs/pactivate.git

We no longer explicitly specify `python3` as the interpreter: since
`pactivate` no longer takes an argument for this we'd instead have to find
the actual intepreter ourselves and create a symlink to it. That doesn't
seem worth the work; it's expected that much Python code won't work on
Python 2 these days anyway.