pylib/cmtconv/Test: Add a -q to make the tests quiet
Output will be completely silent if the tests pass, and maybe even if the
tests fail. This is intended for calling by other test scripts which should
check the result code and generate a failure message.

(If it does fail, the developer is expected to run the script again without
-q to find the error.)