cmtconv..mb6885: Improve "bad magic" error messages for binary files
For the "bad magic" error messages we now print a `$` in front of hex
output and give the `block_num` of the bad block.

This works only for binary files. For BASIC files `block_num` is always
$01. The correct solution here is to build a function that checks the file
type and returns `basic_block_num` or `block_num` as appropriate for that
type. To distinguish the two file types we could probably use `B$nn` and
`S$nn` as the human-readable block number so we know which one is being
returned. (These variables should probably also be renamed to `s_block_num`
and `b_block_num`.)