README: Add notes on parecord to handle saving from a microcomputer
This may not be entirely correct since I've not yet got it working in my
testing with the Basic Master Jr., though it does seem close. (The saves
sound ok but cmtconv will not read them, and when trying to LOAD the saves
them I was at one point seeing the records match properly on the screen.
However, it never actually loaded and now even that is not working.)