cmtconv.audio: Better pulse detection for CMT audio
When reading with only H/L sense we couldn't detect two H or two L pulses
separated by silence, and also we couldn't detect a terminal L followed by
silence.

We now detect H/M/L and use this to detect edges, but are still using just H/L
for the pulse sense, which needs to be fixed.