bastok.msx2: Clarify values for specially treated tokens
For DATA, REM, ELSE and = we need the token value to do direct matches on
it for special handling. We got this out of the TOKENS list using
toklastbyte() because this happened to handle the case for extra-special
ELSE, where we need the token byte after the leading `:`, while not
affecting any of the single-byte tokens. But this was confusing, and won't
work for even more special tokens such as the single-quote alternative to
REM which is three bytes long.

So we replace toklastbyte() with tokbytes() which does the same lookup but
returns the entire sequence of bytes for that token, letting the caller
deal with getting out whichever byte it's looking for.

tokbytes() also now checks to ensure that we have only one token entry for
the given keyword and raises an internal error if we don't. (Previously it
just silently ignored any extra token entries, which could hide typos and
other errors.)

We also change the TOKENS entry for ELSE to use `:` instead of `\x3A`
since that's easier to read.