bin/msx-dasm: Run z80dasm on contents of an MSX-BASIC "BLOAD" file
Machine-language files loaded by MSX-BASIC via BLOAD have a header
giving the address at which it's to be loaded and the entry point.
This strips off that header and passes the rest of the file to
z80dasm, along with an --origin option giving the load address from
the header.

As with the original, this still looks for disassembler info and block
files in info/ directory at the same level as the bin/ directory; this
needs to be fixed somehow.

This is copied from commit c74cf28 of the Izumic Ballade RE project:
  https://gitlab.com/retroabandon/izumic-ballade