bastok.msx2: Clarify parser external/internal APIs
The few functions that most users will use are now at the top; this is
just __init__(), detokenized(), the exception objects and pstate().

External users now call detokenized() which is lazy: if no output has
been generated parse_tline() (ex-detokenize()) is called to do the
parsing; either way the output is then returned.

The old top-level parsing routine, detokenize(), is now renamed to
parse_tline() which guarantees a fresh parse of the current tline
value. This now also generates the line number itself rather than
reset() doing it, since a reset should not then start a little bit of
the parse, especially when external users of the lower-level parse
routines might be resetting and re-parsing.