bin/detok: Add a functional test and fix a bug
When we changed some of the names in the Detokenizer API a while back
we didn't change the call in bin/detok, in part because we didn't have
a test for it.

We add a simple tokenised program `programs/simple.bas` and the
detokenised output `programs/simple.ba`, run bin/detok on the former
and compare it with the latter, which demonstrated the bug (and now
shows that the bug is fixed).