Test,programs/: Test compact detokenisation; better error output
We now detokenise twice, one in "regular" mode that produces compact
output matching the input, and once producing expanded output as before.

This also now does the detokenisation and the comparison of the
results in two separate steps so that we don't get a huge diff after
the exception if detok fails to produce any output.