msx2: Do not expand spaces where we already have them
The expansion of detokenised code was introducing extra space where it
wasn't necessary. Though it did the right thing with `IFX` → `IF X`, it
would also change the already-fine `IF X` → `IF  X` (with two spaces
between the IF and the X).

We fix this by making expand() add a space only when there's not been
a preceeding space and the next character we'll parse is not a space.