Test: Set default testpaths so args w/o explicit paths don't search tmp/
I put a symlink to a clone of the frieve-msxfan-re repo in tmp/. This
caused no harm when running `Test` without arguments, because with no
args it was adding `pylib/` to the arguments. But with `-v` it did not
do so, which caused pytest to do discovery in tmp/frieve-msxfan-re as
well, which caused very mysterious errors and much confusion.

The correct way to deal with this is to set the `testpaths`
ini-option, which will be used if no paths are explicitly specified,
and not add any paths to the pytest command line except what the
developer who ran `Test` gave. Since we don't use an ini file, it's
easier just to add `-o testpaths=pylib/` to the command line than to
create one. This can be overridden by the developer with an additional
`-o` option if she chooses.

WARNING: The `-o testpaths=pyib/` option will override the ini file
if we ever create one, so if we do that we should move that option
to the ini file.