bastok.msx2re: Add 'remain' parameter to digits() test
We probably should have been doing this anyway, but it's been further
inspired by the understanding that you can follow digits with a `%!#`
type indicator _or_ a `[de]NN` exponent, but never both. If you use
`%!#` after an exponent, it's considered not part of the number to be
parsed but instead is following text.