bastok.msx2re match_digits(): now returns len instead of consuming
match_digits() shouldn't consume what it's read because we've not
completed parsing yet. So pass the number of chars we matched back
to the caller, and let it decide if parsing was successful and it
should consume or if it failed and it should leave the text for
something else to try or the error routine to print.