bin/analyze-cmt: Change --bitstream to allow line or char mode
'line' mode is the previous mode, with one bit per line, including timing
information and the like. This is readable as CSV.

'char' mode prints just the `0`, `1` and `x` (unknown) bits in a stream
without line breaks, which is useful for searching for specific patterns in
the input to find the byte and framing format. (We suggest saving a file
with `REM SSSSSSSSS` in it and then searching the output for `01010011` or
the reverse, `11001010`, which will usually make it instantly clear what
the framing is.

Reviewed-by: Curt J. Sampson <cjs@cynic.net>