bin/basdump: Now use -a to print addrs, otherwise line numbers at left
This reduces irrelevant diffs from lines having their address shifted
due to missing or extra code in the line. You can still use -a to
get the old format (albeit now with a couple of extra spaces at
the start to keep both formats aligned the same).

The code is getting rather messy and needs a re-think.