bastok.tlines: Make txttab second param and optional; improve orig_text
It's not actually optional at the moment because it won't yet work
if txttab is not provided, but we're committing the interface change
separately from the support for automatically determining txttab.

We also now set orig_text in parsetext() so that it's set on any
parse when we have no lines. Thus, we can start with an empty
TLines, or use the new clearlines() to clear it, and still have
an orig_text. When we already have lines and parsetext() is merging
with them, we retain the old orig_text. The new text that was merged
is lost, but it's not clear what we can sensibly do otherwise.

This also stats to add constants for txttab values for various systems
(TXTTAB_PET and so on), and improves the comments.

This was originally part of a plan to use this class to read a BASIC
file for bin/basdump, but it's not clear if it's really suited for
that since this wants to throw away and regenerate the line offsets
and basdump should not do that but dump the original ones. But there
is a lot of shared code, so we'll see where this goes.