bin/detok: --binary now implies --dos-text; add functional test
There's no reason we can think of that you'd want to do a save
without charset conversion but in Unix form, since the files will
only have the correct characters displayed on an MSX (or
emulator) itself, which always uses DOS format files. So to help
avoid problems with an MSX machine misinterpreting the output
files, we always enable --dos-text when the --binary option is
given, thus producing the same output as SAVE "...",A produces.

The functional test is programs/binchars.bas and .baa, which are
are binary and ASCII saves from a Japanese MSX machine of a
program containing:
- hiragana あいうえお, which is encoded in single bytes with the
  high bit set, and
- kanji 年月日, which are characters in the low $20 bytes of the
  the character set and thus encoded as two bytes with the $01
  prefix.