BROKEN bin/retok: Add --binary mode
Note you need to temporarily `rm programs/43*` to get past that failure
to the --binary tests.

This doesn't work yet because the higher-level (MSX) parser is trying to
match input against `str`s, but the core PState's input is a `bytes`.
So what we need to do is move all matching into the core parser, and
when the input is not `str` but bytes, the match() function (which should
consume if there's a match?) should convert its "str to match" parameter
to a `bytes` using the ASCII codec before it tries to match.