bastok.parser.{parser,msx2re}: encode_{elem,seq}(); rename number
We were not properly dealing with element vs. sequence types when
they were different; extract routines for this and have all routines
that needed the codec use those routines.

msx2re: The old *digits() routines that parse a BASIC format number
are now renamed *number to avoid confusion with Parser.digit() and
.digits().