bin/basdump,tlines: Start BASFile class for filetype detection
Add a functional test for bin/basdump.

The bastok.tlines.BASFile class checks the header at the start of the file
and separates it from the BASIC `txttab` data. (Currently it knows only
about the MSX-BASIC header.) basdump now uses this to extract the header.

We now always prefix the header line with "HEADER," instead of switching to
an address in address mode, because the header doesn't actually have an
address in memory.