bin/basdump: Large rework out of output format
We now place metadata (next address pointer, lineno) on a separate
line from the data, and generally improve things to make them easier
to read and scan.

We also now have separate programs/basdump.* for the functional test
so that we can better see how it works with various line lengths, and
don't have to change this when data for other tests needs to change.