bastok.msx2re string_literal(): Fix parsing of bytes input
We were matching peek() against DQUOTE, the string '"', but it returns
the input element type, and thus we didn't match b'"'. We fix this by
using Parser.string() to do the match, which handles converting our
`str` DQUOTE to the input type, whatever it is.

Though we are always generating a `bytes`, we were also generating a
`str` to return, which is more difficult when our input is not a
`str`. But we never actually used the return value anyway, so now we
just return `True` if we parsed and generated a string literal; we can
change this later if we ever have anything that needs to know what was
generated (which seems unlikely).

A unit test test_string_literal_bytes() was added, as well as similar
test data to the functional test `programs/binchars.*`.

string_literal() still uses char(p) for generation because it must
generate MSX-BASIC-encoded output; we still need to work out the
right place to deal with the two different encodings, perhaps in
the Parser class.