Test,programs/: Add .ba2 files for alternate forms of input
Like the MSX-BASIC interpreter itself, bastok should correctly read
things that MSX-BASIC and detok don't generate, such as `&H` for `&H0`
and `1.0` for `1!`. The new .ba2 files give a home for those where
needed and, if the .ba2 exists, it's retokenised and tested against
the original .bas file.

We now in the fuctional tests ensure that `&H` and `&O` work.