bastok: Rework &B… to have bytes test and be more clean
We're still having issues with encodings here. We always want to generate
`bytes`; the question is how to deal with the conversion of whatever our
input is to it, if necessary, and of course what we return.

This current thing where we read `bytes`, convert to `str`, and then
convert back when we generate is ridiculous.

Perhaps parsing routines want to be returning a _pair_ of the original
input read and the input converted to `str`.