FAILING show that we should be detokenising nDm for doubles
MS-BASIC does _not_ do this when we do an ASCII save, or even when it lists
the line. Thus, it can't round-trip 6 digit or less significands that have
been forced to double there's no way in its ASCII/list output to see that
it's listing a double instead of a single. (It's not clear why it doesn't
just encode it as a single anyway.)

But we want to be able to preserve such things in a round trip which means
that we should be detokenising differently from MS-BASIC and printing nDm
instead of nEm when we have a double with a significand short enough to fit
into a single. In fact, since we're being different anyway, we should
_always_ detokenise doubles as nDm (even >6 digit significands) because
this will help humans read it more accurately.