bastok.detok.msx2: Change doubles from nEm to nDm format
MS-BASIC, for both LIST and SAVE "…",A, always uses nEm format when it
prints floating point numbers in the ranges that require an exponent. Since
these don't (and can't) have a `!` or `#` trailer, that means that doubles
with 6 sig digs or less don't round-trip: they're read back in (e.g., via
LOAD of an ASCII save) as single-precision. (The nEm format when read
generates single precision for ≤6 sig digs, and double otherwise.)

We want both to indicate what the original tokenised save used and to
maintain round-trip compatibility, so when printing in exponent form a
number tokenised as double, we now always use nDm format. This also makes
it easier for humans to quickly see which it was.