FIXUP Test whexdiff(): Ensure that dump files have different names
The dump file was named after just the name (without path) of the file
being dumped so that if we are comparing, e.g., programs/foo.bas and
.build/.../foo.bas it would dump both to foo.bas.dump. Change it to use
foo.bas.exp.dump and foo.bas.act.dump instead to make sure that a separate
dump file is used for expected and actual inputs.

We didn't notice this problem before because at least some of the callers
were comparing, e.g., .bas and .bas2 files.