bastok.retok.msx2: Parse variable names
We were parsing things that could be variable names as straight
"pass-through" characters, but that would turn a variable name such as
'A1' into a variable 'A' followed by an encoded integer 1, rather than
an ASCII 'A1'.

WARNING: This does not translate lower-case variable names into upper
case, which is probably necessary because that's what MS-BASIC does.

WARNING: This exposed a problem in the detokeniser where it was not
parsing variable names properly, adding a space in `PAGE1` to make it
`PAGE 1` even though it does not know that this is not a variable
name.