bastok.retok.msx2: Test lineno is forced; document a difference
Even though "123!" looks like a short float, it's actually tokenised as
lineno(123) "!" after a GOTO or similar; the new functional test code
confirms this.

Also document how we don't bother to implement some weird behaviour of
MS-BASIC. It will successfully return and continue from
"GOSUB123!9876543210:...", skipping over the 9876543210 that it will have
encoded as a sequence of lineno(9876), lineno(54321), lineno(0).