FIXUP bastok.detok.msx2: All TOKTAB entries now have flags
Previously we had a 2-tuple for entries "without" flags, and a 3-tuple
for entries with any flag set, which was obviously dumb. Now everything's
a 3-tuple and no-flags-set is a 0 value in the flags field.

This lets us use pattern matching to assign names to the fields
in many instances, rather than using an index.

We also now have the flag bitfield indexes as attributes on TOKFLAGS
so that it's clear to clients who owns these constants and they need
not import extra stuff.