FIXUP bastok.detok.msx2: Fix broken "fix" of non-broken sort of DETOKENS
The "fix" in commit cea501c was actually breaking things because we were
confused about what we were doing with DETOKENS (which is parsing the
binary tokens, not the keywords). The tests couldn't show this because
there are no proper prefixes in the actual MSX-BASIC TOKENS table, so we
add a test TOKENS that does have a proper prefix to show that we're now
working properly (again).

Reviewed-by: Stuart Croy <stuartcroy@mac.com>