Test,programs/*: Rework to better test retok --squeeze
This removes a bunch of duplicate stuff in the programs/[0-9][0-9]*
loop and changes it so that the original source (*.bas) is always
fully squeezed, allowing us to test non-squeeze retok using the .ba0
(albeit without testing that spaces get passed through properly) and
squeeze retok with the .ba1. More details about the
programs/[0-9][0-9]* loop are now documented in programs/README.md.

We should probably add a separate test (noted in basrev.md) that confirms
non-expand detok and non-squeeze retok preserves any strange numbers of
spaces that happen to be there.