BROKEN bastok.retok.msx2 linenum(): Show broken pending parse failure
linenum() consumes (pendingly) the '-' that we expect to start a
negative number but then fails when the following character is not a
digit. But this is never rolled back, our pending parse point is still
at the number instead of the '-' and any following p.commit() will
leave our parse point after the '-' that has never been correctly
parsed.