pyproject.toml: Initial Python PIPy configuration (v0.0.0)
- pyproject.toml has general metadata, setuptools as the backend, and uses
  all import packages under pylib/.
- Update `requirements.txt` updated to include `build`, which provides the
  setuptools back-end we need for this project.
- .gitignore ignores the output files from `pyproject-build` (the
  command-line program from the `build` package).

This will allow us to use `pip install -e .` (in the virtualenv) to do an
"editable" install, redirecting to the source files in pylib via a .pth
file (lib/python*/site-packages/__editable__.r8format-0.0.0.pth), giving us
the r8format libraries in our virtual environment. This isn't necessary for
the programs and tests in this repo (which add it themselves), but is
necessary to make libraries such as binary.* available to t8dev or other
tools that need them.

We don't yet do the `pip install -e`, nor have we set up the binaries in
bin/* to be installed into the virtualenv. (Nor is this suitable to be
uploaded to PyPI for these reasons and because we don't yet have a version
number.)