Test: pip install -e r8format
This makes the libraries usable by other code using the virtualenv without
them having to set a path to r8format/pylib/.

We install (or re-install) r8format only if it's not already installed as
editable. This saves time if it's already properly installed, as the
install takes several seconds, and also forces a re-install if it wasn't
installed as editable.