doc/,pyproject: long_description and updates for docs
long_description typically contains the contents of the REAMDE file and
long_description_content_type defaults to `text/x-rst` or whatever. The
documentation only suggests setting up setup.py code to load these, which
is obviously wrong for pyproject.toml builds.

As it turns out, you can set `readme` in pyproject.toml and build tool will
automatically deal with this for you, including setting the content type
based on the extension. So we set `readme = 'README.md'`, long_description
will be filled with that, and the content type will automatically be set to
`text/markdown`.