pyproject: Add top-level pytest_pt.py to distribution package
After much debugging we determined that we're not using setuptools
automatic discovery (of either form) but custom discovery, because we have
a [tool.setuptools.packages.find] table. The standard `find` routine
doesn't seem to work there as the one for automic discovery in src-layout
mode: it accepts everything except top-level modules in files.

We work around this by using the (pratically uncodumented) `py-modules`
array which seems to add module names (not filenames) in it to the
distribution package, not affecting anything else that's been found.

There's considerably more information at this Stack Overflow answer:
https://stackoverflow.com/a/78382535/107294

Reviewed-by: Curt J. Sampson <cjs@cynic.net>