binary.tool.asl: Better error msg when failing to parse values
When we fail to parse a falue from a list or map file, we now print the
failed value, type we expected, and the whole line (including the name) for
context.

This was prompted by the parser failing on DATE and TIME symbol data when
`LC_TIME=ja_JP.UTF-8`. That causes the output to include Japanese
characters (e.g. 2024年7月14日), which is fine in the listing file, but
produces escaped `\nnn` chars with negative values (e.g., `\-27`) when the
symbol values are printed. (That should be `\229\.)

This doesn't change how we handle this (we still fail to parse), but now
it's clear where the problem is and what workaround (set your locale to
ASCII-only) is necessary until the ASL bug gets fixed.