binary.romimage: Do not catch HTTPError
The code catching HTTPError (e.g., when we can't download from a URL we've
been given) and trying to print an error message was a hangover from when
this was in t8dev. (And failing, because we don't have an err() function
here.) Just let the HTTPError propagate up and the caller can deal with it
or not as it likes.