binary.romimage patches(): Fix breakage when >1 patchspec matches
We were immediately removing each patchspec from patchspecs as we used
it, but still using the original indices we'd generated rather than
the new indices. This caused us a) to use the wrong patchspec for the
next match, and b) delete the wrong one (or die when trying to delete
past the end of the list).