WIP binary → r8bin: Change module name to help avoid collisions
While our top-level `bastok` and `cmtconv` module names appear to be unique
on PyPI and are unlikely to be used for something else, `binary` is used by
a package ("binary") on PyPI and is rather generic for a top-level module
name anyway. So we rename this top-level module to `r8bin`, which appears
to be unique on PyPi and is unlikely to be used by anybody else.

We considered putting everything under an `r8format` top-level module
(r8format.bastok, r8format.binary, r8format.cmtconv)...

XXX so why not use the one top-level module? Kinda makes it more clear
for dependencies, too, when you look at what they're importing.

XXX Update Changelog, bump version to 1.x