binary.cli.p2b: New program to generate Apple II and KC85 program files
The `p2b` program reads the .p files generated by ASL and generates a
binary file with a header with the appropriate load address, length
and (where necessary) entry address. Adding other platforms to this is
easy, and we should consider adding some sort of specification language
to do this for arbitrary runnable program headers. (But possibly it's
so easy to add new machine entries that that's not worthwhile.)